refactoring utils/ folder
