--- src/rarelink/utils/schema_processing/__init__.py
@@ -1,19 +1,0 @@
-"""
-Schemas Processing Submodule
-
--> This submodule contains all functions to allow processing of specific data 
-schemas to required output formats.
-
-Exports:
-- 
--
-- 
-"""
-
-from .redcap_to_linkml import redcap_to_linkml
-from .linkml_to_redcap import linkml_to_redcap
-
-__all__ = [
-    "redcap_to_linkml",
-    "linkml_to_redcap"
-]
