--- src/rarelink/utils/validation/__init__.py
@@ -1,12 +1,0 @@
-"""
-This module contains all RareLink validation tools and functionalities.
-
-"""
-
-from .linkml_data import validate_linkml_data
-from .hgvs import validate_and_encode_hgvs
-
-__all__ = [
-    "validate_linkml_data",
-    "validate_and_encode_hgvs"
-]
